move benchmarking periphery into a module
