+++ runtime/src/lib.rs
@@ -159,8 +159,8 @@ use frame_support::traits::{Everything, ExistenceRequirement, Get, PrivilegeCmp,
 #[cfg(all(feature = "private-net", feature = "wip"))] // order-book
 pub use qa_tools;
 pub use {
-    assets, eth_bridge, frame_system, multicollateral_bonding_curve_pool, order_book, trading_pair,
-    xst,
+    assets, eth_bridge, frame_system, multicollateral_bonding_curve_pool, order_book,
+    order_book_benchmarking, trading_pair, xst,
 };
 
 /// An index to a block.
